requirements.txt
openai
SpeechRecognition
python-dotenv
pyaudio
gTTS
pyautogui
pydub
pygame
pyttsx3